Refactor and enhance audio processing and configuration handling
- Updated generator_capture_read to ensure frequency delta is within valid range.
- Changed stopped flags in WASAPI backend to atomic for thread safety.
- Enhanced filter and mixer configuration equality checks to include additional parameters.
- Improved error handling in filter and processor configuration parsing.
- Added support for samplerate and channels in JSON configuration parsing.
- Introduced FILTER_TYPE_INVALID and PROCESSOR_TYPE_INVALID to handle unknown types.
- Fixed memory allocation issues in device capabilities retrieval.
- Refactored async resampler to improve handling of zero output frames.
- Enhanced lock-free ring buffer to include sequence tracking for better synchronization.
- Added null checks and error reporting in various functions to improve robustness.

+++ FFT/bluestein_fft.c
@@ -121,8 +121,8 @@ bluestein_fft_t* bluestein_fft_create(size_t n, config_error_t* err) {
     int min_k = (f == 1) ? 4 : 3;
     double target = (double)min_l / (double)f;
     int k = min_k;
-    while ((1 << k) < target) k++;
-    size_t m_val = (size_t)f * (1 << k);
+    while (((uint64_t)1 << k) < target) k++;
+    size_t m_val = (size_t)f * ((size_t)1 << k);
     if (m_val < best_m) best_m = m_val;
   }
   size_t m = best_m;
